+++ python/coinbase-agentkit/pyproject.toml
@@ -15,6 +15,7 @@ pydantic = "^2.0"
 web3 = "^7.6.0"
 python-dotenv = "^1.0.1"
 requests = "^2.31.0"
+paramiko = "^3.5.1"
 
 [tool.poetry.group.dev.dependencies]
 ruff = "^0.7.1"
