--- python/coinbase-agentkit/tests/action_providers/hyperboliclabs/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for the Hyperbolic action provider."""
